Add a config header which contains the program version.
Since this config.hpp is generated in the makefile, the dependencies
must be generated with the -MG flag to avoid a makefile error.  But
this -MG flag only works if -M or -MM is used.  But if -M or -MM is
used, the code is not compiled.  Therefore a separate rule to generate
the dependencies (*.d) without compilation is introduced.  These
dependency files are then included and processed, which leads to
compilation.